posix-cpu-timers: Split out posix-timers_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

Trimming down sched.h dependencies: we don't want to include more than
the base types.

Cc: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 53d31ba842d9cc391032d051a210c3c9941f1529)
